rien de bien fou

+++ routes/web.php
@@ -22,16 +22,16 @@ Route::get('/', function () {
 Route::get('/register', [RegisterController::class, 'showRegistrationForm'])->name('register');
 Route::post('/register', [RegisterController::class, 'register']);
 
-Route::get('/search', [SearchController::class, 'search'])->name('search');
-Route::get('/places', [PlaceController::class, 'index'])->name('places.index');
-Route::get('/places/{place}', [PlaceController::class, 'show'])->name('places.show');
+Route::get('/search', [SearchController::class, 'search'])->name('search')->middleware('auth');
+Route::get('/places', [PlaceController::class, 'index'])->name('places.index')->middleware('auth');
+Route::get('/places/{place}', [PlaceController::class, 'show'])->name('places.show')->middleware('auth');
 Route::get('/', [VisitorController::class, 'index']);
-Route::get('/places/create', [PlaceController::class, 'create'])->name('places.create');
-Route::post('/places', [PlaceController::class, 'store'])->name('places.store');
+Route::get('/places/create', [PlaceController::class, 'create'])->name('places.create')->middleware('auth');
+Route::post('/places', [PlaceController::class, 'store'])->name('places.store')->middleware('auth');
 
-Route::get('/places/{place}/edit', [PlaceController::class, 'edit'])->name('places.edit');
-Route::put('/places/{place}', [PlaceController::class, 'update'])->name('places.update');
-Route::delete('/places/{place}', [PlaceController::class, 'destroy'])->name('places.destroy');
+Route::get('/places/{place}/edit', [PlaceController::class, 'edit'])->name('places.edit')->middleware('auth');
+Route::put('/places/{place}', [PlaceController::class, 'update'])->name('places.update')->middleware('auth');
+Route::delete('/places/{place}', [PlaceController::class, 'destroy'])->name('places.destroy')->middleware('auth');
 
 
 Route::middleware('auth')->group(function () {
@@ -47,12 +47,10 @@ Route::get('/my-reservations', [ReservationController::class, 'myReservations'])
     ->middleware('auth')
     ->name('reservations.index');
 
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-
 Auth::routes();
 
 
-Route::get('/profile', [ProfileController::class, 'show'])->name('profile');
+Route::get('/profile', [ProfileController::class, 'show'])->name('profile')->middleware('auth');
 
 
 Route::get('/report', [ReportController::class, 'index'])->name('report.index');
@@ -68,21 +66,21 @@ Route::middleware(['auth'])->group(function () {
     Route::resource('object_co', Object_CoController::class);
 });
 
-Route::resource('events', EventController::class);
+Route::resource('events', EventController::class)->middleware('auth');
 
-Route::get('/events', [EventController::class, 'index'])->name('events.index');
-Route::get('/events/create', [EventController::class, 'create'])->name('events.create');
-Route::post('/events', [EventController::class, 'store'])->name('events.store');
+Route::get('/events', [EventController::class, 'index'])->name('events.index')->middleware('auth');
+Route::get('/events/create', [EventController::class, 'create'])->name('events.create')->middleware('auth');
+Route::post('/events', [EventController::class, 'store'])->name('events.store')->middleware('auth');
 
-Route::get('/events/{event}/edit', [EventController::class, 'edit'])->name('events.edit');
-Route::put('/events/{event}', [EventController::class, 'update'])->name('events.update');
-Route::delete('/events/{event}', [EventController::class, 'destroy'])->name('events.destroy');
+Route::get('/events/{event}/edit', [EventController::class, 'edit'])->name('events.edit')->middleware('auth');
+Route::put('/events/{event}', [EventController::class, 'update'])->name('events.update')->middleware('auth');
+Route::delete('/events/{event}', [EventController::class, 'destroy'])->name('events.destroy')->middleware('auth');
 
-Route::get('/profile/{user}', [ProfileController::class, 'show'])->name('profile.show'); // Page publique
+Route::get('/profile/{user}', [ProfileController::class, 'show'])->name('profile.show')->middleware('auth'); // Page publique
 Route::middleware(['auth'])->group(function () {
     Route::get('/profile/{user}/edit', [ProfileController::class, 'edit'])->name('profile.edit'); // Page privée
     Route::put('/profile/{user}', [ProfileController::class, 'update'])->name('profile.update'); // Mise à jour
 });
 
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
-Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show');
+Route::get('/users', [UserController::class, 'index'])->name('users.index')->middleware('auth');
+Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show')->middleware('auth');
